More renaming to follow python conventions
