+++ .github/workflows/windows.yml
@@ -28,9 +28,9 @@ jobs:
       working-directory: build
       run: nmake run_tests
 
-    - name: Build libPyTweedledum
+    - name: Build python lib
       working-directory: build
-      run: nmake libPyTweedledum
+      run: nmake _tweedledum
 
     # I will not run the tests unless my hands on a windows machine
     # Unfortunately, I cannot reproduce the errors :(
