+++ python/tweedledum/__init__.py
@@ -2,5 +2,5 @@
 # Part of Tweedledum Project.  This file is distributed under the MIT License.
 # See accompanying file /LICENSE for details.
 #-------------------------------------------------------------------------------
-from .libPyTweedledum import *
+from ._tweedledum import *
 from .bool_function_compiler import *
